Add in .panthoenrc & new binaries;
The .panthoenrc file is going to be the main configuration file for all pantheon projects on the user's computer. It can be found at the $HOME dir level, or the current project's root dir.

The new binaries are some simple commander boilerplate using RSVP to handle some async file calls. I want to split this up into modules so hopefully I can consolidate these. For now, it is a succotash of old and new scripts with the updated commander-command syntax.