Making some progress on media command;
Brought in the old bash script as comments on the bottom of paw-media.
I'm going to use the npm rsync module I found to perform the same
thing. I need to do error handling for parsing JSON. I believe it
fails silently and will be difficult to debug when there is a parsing
issue in the .pantheonrc file.